gardenbook_chat_api/requirements.txt
fastapi
uvicorn
mcp-python
langchain-core
langchain-anthropic
langchain-ollama
langgraph
langchain-mcp-adapters
python-dotenv
pymongo 